escher_vendor_brain_v0_1/requirements.txt
pandas==2.2.2
pyyaml==6.0.2
requests==2.32.3
tqdm==4.66.4
rapidfuzz==3.9.6
